properly implemented force chunks + made a ClientTardisUtil to avoid errors
