+++ src/main/java/mdteam/ait/client/AITModClient.java
@@ -9,6 +9,7 @@ import mdteam.ait.client.renderers.entities.ControlEntityRenderer;
 import mdteam.ait.client.renderers.entities.FallingTardisRenderer;
 import mdteam.ait.client.renderers.exteriors.ExteriorRenderer;
 import mdteam.ait.client.screens.MonitorScreen;
+import mdteam.ait.client.util.ClientTardisUtil;
 import mdteam.ait.core.AITBlockEntityTypes;
 import mdteam.ait.core.AITEntityTypes;
 import mdteam.ait.core.AITItems;
@@ -211,7 +212,7 @@ public class AITModClient implements ClientModInitializer {
                             if (!tag.contains("tardis")) {
                                 return;
                             }
-                            TardisUtil.snapToOpenDoors(UUID.fromString(tag.getString("tardis")));
+                            ClientTardisUtil.snapToOpenDoors(UUID.fromString(tag.getString("tardis")));
                         }
                     }
                 } else {
